modify knex migrations

+++ server/db/migrations/20250102094949_create_files_table.js
@@ -3,17 +3,22 @@
  * @returns { Promise<void> }
  */
 exports.up = function (knex) {
-  return knex.schema.createTable("files", (table) => {
-    table.increments("id");
-    table.string("filename").notNullable();
-    table.string("s3_key").notNullable();
-    table
-      .integer("user_id")
-      .unsigned()
-      .references("id")
-      .inTable("users")
-      .onDelete("CASCADE");
-    table.timestamps(true, true);
+  knex.schema.hasTable("files").then((exists) => {
+    if (!exists) {
+      return knex.schema.createTable("files", (table) => {
+        table.increments("id").primary();
+        table.string("filename").notNullable();
+        table.string("s3_key").notNullable();
+        table
+          .integer("user_id")
+          .unsigned()
+          .references("id")
+          .inTable("users")
+          .onDelete("CASCADE");
+        table.timestamps(true, true);
+      });
+    }
+    return Promise.resolve()
   });
 };
